in this changeset, blocks defer to sends to print
formatting send chains is easier when the block defers to the send to
print the send and the chain as a doc within the send for breaking up
the args, block begin, and send method optionally (conditional group).
This is largely WIP and I did not obtain _all_ of the results I wanted.
Namely, a long chain or methods with args with a block at the end will
break the last method, but the first few will not be broken and exceed
line length. I can only surmise the doc printing algorithm reaches a
depth in which it does not think it can recurse deeper to obtain any
results by breaking further groups, so it stops and does not recurse
further. This is speculation, I can't explain why the chain doesn't
break further.